correctly reserve capacity for values
before i was initializing the vector with 100000 blank entries, and any
further elements would be added on top of that. presumably it was still
working becauuse if you initialize the vector like that it reserves more
capacity than just the 100000. We could probably get away with way less
reserved, but it doesn't seem necessary for testing