in Value struct: Value * -> ValueRef
changed all instances of Value * in the Value struct itself into
ValueRefs. Then changed anything that needed it to call VPtoVR() or
VRtoVP(). This is not a clever change. This is a minimal effort change.
In Phase II, much of this will get cleaned up.

Additionally, anything remotely complicated crashes because sometimes
(looking at you Exprs) values get stored outside the vector, so
ValueRefs pointing at them are invalid. This will be fixed in a later
commit.