in Expr*: Value -> ValueRef
Values must all be in the EvalState's values vector in order to be
referenceable with ValueRefs. As such, we can't store them as part of an
Expr. So those Exprs have to store ValueRefs instead. This does increase
memory usage by 4 bytes, since now we have a Value allocated and *also*
a ValueRef, where before we had only a Value. This also adds a layer of
indirection, which could cause more cache misses.

Certain things now work (e.g evaluating [1]) but other things still
don't, because there are still values allocated outside of the vector.