for SymbolTable: Value -> ValueRef
This one was quite a bit more complex. symbol-table.hh has grown to be a
convoluted monster over the years. As part of this change I tried to add
good documentation comments and simplify the implementation. I also
intend to rename the classes to things that make more sense for their
current usage, but I want to do that in a separate commit so that it's
clearer what each one is doing.

This commit removes the separate storage of SymbolValues inside the
SymbolTable. All values must be stored in EvalState's values vector!

The SymbolTable had already been quite well optimized though, so making
this change required moving things around to avoid adding any bytes.

I noticed the use of ChunkedVector for Values, and will need to consider
using that instead of vector. That comes later when we can measure the
performance impact.