rename Symbol -> SymbolRef and SymbolStr -> Symbol
this is a bit confusing if you're used to the old naming scheme, since
Symbol refers to something different now. However, I think this naming
scheme matches more closely how we're actually using these structs now.

SymbolData: underlying data
Symbol: reference to SymbolData. This is the layer at which things are
deduplicated
SymbolRef: an alternative to Symbol that is 4 bytes but requires
EvalState to fetch the data, like ValueRef. Well, exactly like
ValueRef...