update comments for SmallTemporaryValueVector
It was always true that you couldn't keep references to these Values
after they left scope. But now we have the added restriction that we
cannot **ever** take ValueRefs of them, since they are not in the
vector.

The good news is, I think this is already the case. I ran a few tests
trying to use self-reference to get one of the Values into some
descendant of itself or another one of them, and it didn't happen.