resolve the vNull situation
vNull (and others) were Values stored directly in the EvalState object.
Now they are ValueRefs pointing to Values stored in the vector. Even
better, we could put them at well-known locations and have them be
constants, but that's a very small optimization and we can deal with it
later if we want to.