get rid of remaining Value **s
I've fully bit the bullet and decided to take all stack-allocated Values
and allocate them in values. I think this will be necessary eventually,
since Value will cease to mean anything outside of that context.
However, I want to keep track of where in the code these used to be (so
it's easy to do something with them later), and how many of them there
are (so we can see the impact).

One idea for how to deal with them is to create a separate value array
for temporary values that then get bump