LazyPosAccessors: Value -> ValueRef
Well really i had to deconstruct the whole thing. The primop values
can't live in a static struct, they msut be in the values vector like
everything else, so i initialized them with the EvalState and got rid of
LazyPosAccessors altogether. I think the properties we want about not
being able to access the primops from anywhere else still hold.