Deal with Values on the stack
We have a bunch of Values on the stack. As things were, they could not
be used anywhere where we needed a ValueRef. There are a couple ways to
deal with this

1. Allocate them in the values vector instead. This is in some ways the
easiest patch. However, I added in some debug statements and it seems
that in a nixos configuration eval there are about 3x more Values
allocated on the stack than on the heap. That's way too many.
Nevertheless, this could have been a stopgap measure on the way to (2)

2. Make a separate vector for values "on the stack", that replicates the
behaviour of the stack (bump allocator). This in theory feels like the
"right" way to do this, but in practice it would be very error prone,
since you have to remember for every function that allocates Values on
the stack to (a) take the stack pointer at the beginning and (b) reset
it at the end. It would also introduce big inefficiencies if you don't
remember to e.g. allocate Values outside your loops. All of these things
are not easily checked and might not produce any obvious problems for a
while if ever.

3. Encode pointers to stack Values in ValueRefs. The upsides of this are
that we can use Values on the stack like normal, with the usual fast
allocation / deallocation and compiler-managed stack that we do now. The
downside is that we have to make a couple of assumptions about the
stack:

   1. Assuming the stack is contiguous in memory.
   2. Assuming the stack is "small" (less than ~1GB in size)

These assumptions seem to hold in most contexts, and I've decided on
balance to go with (3) for now, but it does make me nervous that this
will break Nix in some obscure context. However, I will say that Boehm
GC also breaks on weird systems for similar reasons.

Ideally, we could get a pointer to the bottom of the stack, but in
practice this is even less portable, so I've implemented it by taking a
pointer to a random spot in the stack and having the offset be signed,
since now our Values could be above or below that point.

As an added bonus change, I've made ValueRefNull 0. There are a bunch of
places already in the code base where we compare to 0 to test whether
ValueRefs are valid, and I think this is also the faster approach. So we
offset Value indices by 1 when encoding them in a ValueRef.