SymbolData must store the string data, even for empty string
We had a special case for the empty string symbol, which would refer
directly to the C++ string "". I'm not sure if this enabled some useful
optimizations, but it also was a problem for correctness, since we're
now using the Value's c_str field as a back-pointer to the SymbolData.
As such c_str must actually point to the string data of the SymbolData,
so we can't get away with this special case anymore.