allocValue() return ValueRef
The next phase has changed slightly from what I intended. making sure we
don't hold a Value * past a call to allocValue() required a bunch of
tricky reasoning about stack variables.

So instead I'm going to go function-by-function and remove all Value &
and Value *s in arguments, return values, and on the stack. This will be
a huge refactor, but it can be done function-by-function, and is fairly
straightforward inshallah. We also would have to do it eventually
anyways, it's just a question of how we get there.