git rid of SmallTemporaryValueVector
There isn't really a way to make it work with the new memory model.
Fortunately, we don't need it. In both cases we were using it, we
actually only needed the data contained in the Value to persist between
loop iterations. In one case, this work had already been done and the
SmallTemporaryValueVector was totally unnecessary. In the other case it
could be replaced with a SmallVector referring only to the ListViews of
those Values.