eval.hh: Value * -> ValueRef
I removed the variant of addConstant() that returned a Value *. It was
only being used from addPrimOp(). I removed the return value from that
as well, as it was being used nowhere.

With the changes I'm making they would have been functions that only
differed in their return type, which is not allowed. As it is, they only
differed in their return type and whether one of their arguments was a
pointer or a reference, which seems like a bad idea... (very easy to
accidentally call the wrong function, as was happening)

Obviously VRtoVP() and VPtoVR() keep their Value * (eventually the
functions will be going away altogether)