introduce addConstant() that operates on (Stack)Values
Change of plans. it turns out the two different addConstant()s were more
than cosmetic. If a Value & was passed in, it would copy it into the
heap. If a Value * was passed in, it would not. This seems a little bit
error prone...

But in any case, I changed it now to take in either a ValueRef or a
Value. I know we're trying to get *rid* of Values, but in this case I
think it's justified. Eventually we will have StackValues that represent
Values only on the stack, and this is actually exactly what we want. If
we pass in a StackValue, we want it to copy it onto the heap. If we pass
in a ValueRef, the assumption is it's already on the heap. We can even
verify that by inspecting the ValueRef.