apply the rest of the changes from my first attempt
I found the bug!

There was a line `if (&v1 == &v2)` that originally operated on Value &s
(thus testing pointer equality), but I changed those to ValueRefs. It
still typechecked, but now we were testing essentially
pointer-to-pointer equality, which is not what we wanted.

I've applied almost all the changes from the first time so that we can
mor easily apply all the rest of the commits from that branch on top.