fix cursed bug in boolean operations
we can't call
VRtoV(foo).mkBar(function(baz))

because it does things in the following order
1. gets the Value &
2. calls function(baz)
3. calls mkBar()

but (2) might call allocValue() and invalidate the Value & before we've
used it.

I'm honestly surprised this is the only place in the code base this
triggers. Surely there are other examples of it? In any case, once
mkBar() functions directly operate on ValueRefs, this won't be an issue.
Or at least the places we have to be careful about it will be much
smaller (only the methdos of ValueRef)