rework SymbolRef slightly
1. get rid of Value::toPtr() and Symbol::valuePtr(). In every case they
were being used to convert from a Symbol back to a ValueRef after we had
just used a ValueRef to get that Symbol.
2. rename toPtr() to ref() in case we ever need it. it does seem like
there should be a function to get from a Symbol to the SymbolRef
associated with it, so there it is.
3. Make SymbolData contain a SymbolRef rather than a ValueRef. It is by
definition a valid SymbolRef; might as well have the types reflect that.