add ValueRef::set() ValueRef::setFromStack() ValueRef::toStack() get rid of EvalState::VRtoV(), EvalState::VRtoVP()
I've kept Values::VRtoV() and Values::VRtoVP(), but this means that
their reach is much more contained.

I've also added a toStack() function that does what VRtoV() did, but is
expected to be used only when you actually want a StackValue.

set() and setFromStack() were the last remaining uses of VRtoV() which
was setting the content behind a ValueRef.