add Value::ref() get rid of EvalState::VPtoVR()
in some ways this is just a renaming, but it carries with it a
difference in mentality (and implementation!) Value::ref() is
exclusively for acting on stack values. it does not act on Value *, and
it wouldn't know what to do with a value in values if it somehow got its
hadns on one. This is important, because in the very near future values
won't contain contiguous Values!