limit Payload (in heap values) to 8 bytes
1. there's no such limitation on stack values, and splitting off extra
data puts things on the heap we have to clean up, which is better to
avoid

Eventually, we should be able to make most variants 8 bytes, at which
point it may be better to simplify the code and unify the two Payloads
again.

2. as of right now this actually *increases* our memory a good deal.
this is because (a) every oversized value takes up an additional 8
bytes, but more importantly (b) every thunk is leaking 16 bytes of
information once it gets transformed into its target type.