special-case ExprBlackHole
it works differently because there aren't actually any Exprs of this
type, we just have one ghost value that we use for its address.

Eventually, when we don't have any need of Expr *s, we should revisit
this and see if it can be simplified. perhaps we can have an
ExprRef::blackHole like we have ExprRef::null