deal with ExprVar specially
ExprInheritFrom is a subtype of ExprVar, which was messing up some
conversions. We would emplace ExprInheritFrom, and then get back an
ExprInheritFromRef. At some point we would convert it to a
ExprInheritFrom *, and call a method on it. This method would resolve to
the method on ExprVar *, which would expect a true ExprVar.

For now I've just manually coded the right conversions. This is messy
and error-prone (it took me several tries to get it right). The good
news is, all of this will get ripped out and the remaining bits
refactored later on. We won't even have methods on ExprFoo *s at all.

That means I'm not going to worry about doing this neatly.