move maybeThunk() from ExprFoo to ExprRef
in this case the code is much simpler if we just implement it in one
function. in theory it is slightly faster in some cases to have it
implemented by ExprRef sub-type, because then we can skip the switch
statement if we already know the type. In practice I cannot imagine this
is matters.